python_feedforward/requirements.txt
python-mnist
numpy
requests
